+Added basic tests for all addressing modes

+++ NESseract/NESseract.Core/Cpu/AddressingModes/IAddressingMode.cs
@@ -2,7 +2,7 @@
 {
    public interface IAddressingMode
    {
-      byte GetValue(CPUMemory memory, CPURegisters registers, byte operand1, byte operand2);
+      ushort GetValue(CPUMemory memory, CPURegisters registers, byte operand1, byte operand2);
 
       string GetSyntax(byte operand1, byte operand2);
    }
