+++ NESseract/NESseract.Core/Cpu/AddressingModes/ZeroPageAddressingMode.cs
@@ -2,7 +2,7 @@
 {
    public class ZeroPageAddressingMode : IAddressingMode
    {
-      public byte GetValue(CPUMemory memory, CPURegisters registers, byte operand1, byte operand2)
+      public ushort GetValue(CPUMemory memory, CPURegisters registers, byte operand1, byte operand2)
       {
          return memory.Memory[operand1];
       }
